# backend/docs-static/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"SlotResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateSlotRequest": {
				"properties": {
					"slotNumber": {
						"type": "string"
					},
					"slotType": {
						"type": "string",
						"enum": [
							"REGULAR",
							"COMPACT",
							"EV",
							"HANDICAP_ACCESSIBLE"
						]
					}
				},
				"required": [
					"slotNumber",
					"slotType"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateSlotRequest": {
				"properties": {
					"status": {
						"type": "string",
						"enum": [
							"AVAILABLE",
							"OCCUPIED",
							"MAINTENANCE"
						]
					},
					"slotType": {
						"type": "string",
						"enum": [
							"REGULAR",
							"COMPACT",
							"EV",
							"HANDICAP_ACCESSIBLE"
						]
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"AutoAssignResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"slot": {
						"properties": {
							"slotType": {
								"type": "string",
								"enum": [
									"REGULAR",
									"COMPACT",
									"EV",
									"HANDICAP_ACCESSIBLE"
								]
							},
							"slotNumber": {
								"type": "string"
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"slotType",
							"slotNumber",
							"id"
						],
						"type": "object"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AutoAssignRequest": {
				"properties": {
					"vehicleType": {
						"type": "string",
						"enum": [
							"CAR",
							"BIKE",
							"EV",
							"HANDICAP_ACCESSIBLE"
						]
					}
				},
				"required": [
					"vehicleType"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"SendOTPResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"otp": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"message",
					"otp"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SendOTPRequest": {
				"properties": {
					"email": {
						"type": "string"
					}
				},
				"required": [
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"VerifyOTPResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"user": {
						"properties": {
							"createdAt": {
								"type": "string",
								"format": "date-time"
							},
							"email": {
								"type": "string"
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"createdAt",
							"email",
							"id"
						],
						"type": "object"
					},
					"token": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"VerifyOTPRequest": {
				"properties": {
					"email": {
						"type": "string"
					},
					"otp": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"otp"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AuthUser": {
				"properties": {
					"id": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"email",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AuthResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"user": {
						"$ref": "#/components/schemas/AuthUser"
					}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"paths": {
...
		"/api/slots": {
			"get": {
				"operationId": "GetSlots",
				"responses": {
					"200": {
						"description": "Slots retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SlotResponse"
								}
							}
						}
					}
				},
				"description": "Get all slots",
				"summary": "Get all parking slots with filters",
				"tags": [
					"Slots"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"AVAILABLE",
								"OCCUPIED",
								"MAINTENANCE"
							]
						}
					},
					{
						"in": "query",
						"name": "type",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"REGULAR",
								"COMPACT",
								"EV",
								"HANDICAP_ACCESSIBLE"
							]
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 50,
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateSlot",
				"responses": {
					"201": {
						"description": "Slot created successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SlotResponse"
								}
							}
						}
					},
					"400": {
						"description": "Invalid request or slot number already exists"
					}
				},
				"description": "Create new slot",
				"summary": "Create a new parking slot",
				"tags": [
					"Slots"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateSlotRequest"
							}
						}
					}
				}
			}
		},
...
		"/api/slots/{slotId}": {
			"get": {
				"operationId": "GetSlotById",
				"responses": {
					"200": {
						"description": "Slot retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SlotResponse"
								}
							}
						}
					},
					"404": {
						"description": "Slot not found"
					}
				},
				"description": "Get slot by ID",
				"summary": "Get specific slot details",
				"tags": [
					"Slots"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "slotId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateSlot",
				"responses": {
					"200": {
						"description": "Slot updated successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SlotResponse"
								}
							}
						}
					},
					"404": {
						"description": "Slot not found"
					}
				},
				"description": "Update slot",
				"summary": "Update slot status or type",
				"tags": [
					"Slots"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "slotId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateSlotRequest"
							}
						}
					}
				}
			}
		},
		"/api/slots/auto-assign": {
			"post": {
				"operationId": "AutoAssignSlot",
				"responses": {
					"200": {
						"description": "Slot assigned successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AutoAssignResponse"
								}
							}
						}
					},
					"404": {
						"description": "No available slots for vehicle type"
					}
				},
				"description": "Auto-assign slot",
				"summary": "Automatically assign best available slot for vehicle type",
				"tags": [
					"Slots"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AutoAssignRequest"
							}
						}
					}
				}
			}
		},
		"/api/slots/{slotId}/maintenance": {
			"post": {
				"operationId": "SetSlotMaintenance",
				"responses": {
					"200": {
						"description": "Slot marked for maintenance",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SlotResponse"
								}
							}
						}
					},
					"400": {
						"description": "Slot is currently occupied"
					}
				},
				"description": "Set slot to maintenance",
				"summary": "Mark slot as under maintenance",
				"tags": [
					"Slots"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "slotId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/slots/{slotId}/release-maintenance": {
			"post": {
				"operationId": "ReleaseSlotMaintenance",
				"responses": {
					"200": {
						"description": "Slot released from maintenance",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SlotResponse"
								}
							}
						}
					},
					"404": {
						"description": "Slot not found"
					}
				},
				"description": "Release slot from maintenance",
				"summary": "Mark slot as available after maintenance",
				"tags": [
					"Slots"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "slotId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/slots/bulk": {
			"post": {
				"operationId": "BulkCreateSlots",
				"responses": {
					"201": {
						"description": "Slots created successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SlotResponse"
								}
							}
						}
					},
					"400": {
						"description": "Invalid request or duplicate slot numbers"
					}
				},
				"description": "Bulk create slots",
				"summary": "Create multiple slots at once",
				"tags": [
					"Slots"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"$ref": "#/components/schemas/CreateSlotRequest"
								},
								"type": "array"
							}
						}
					}
				}
			}
		},
...
		"/api/auth/send-register-otp": {
			"post": {
				"operationId": "SendRegisterOTP",
				"responses": {
					"200": {
						"description": "OTP sent successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SendOTPResponse"
								}
							}
						}
					},
					"400": {
						"description": "Invalid request or user already exists"
					}
				},
				"description": "Send OTP for user registration",
				"summary": "Send OTP to email for registration",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SendOTPRequest"
							}
						}
					}
				}
			}
		},
		"/api/auth/verify-register-otp": {
			"post": {
				"operationId": "VerifyRegisterOTP",
				"responses": {
					"200": {
						"description": "Registration successful",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/VerifyOTPResponse"
								}
							}
						}
					},
					"400": {
						"description": "Invalid OTP or registration failed"
					}
				},
				"description": "Verify OTP and create user account",
				"summary": "Verify registration OTP and create account",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/VerifyOTPRequest"
							}
						}
					}
				}
			}
		},
		"/api/auth/send-login-otp": {
			"post": {
				"operationId": "SendLoginOTP",
				"responses": {
					"200": {
						"description": "OTP sent successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SendOTPResponse"
								}
							}
						}
					},
					"400": {
						"description": "Invalid request or user not found"
					}
				},
				"description": "Send OTP for user login",
				"summary": "Send OTP to email for login",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SendOTPRequest"
							}
						}
					}
				}
			}
		},
		"/api/auth/verify-login-otp": {
			"post": {
				"operationId": "VerifyLoginOTP",
				"responses": {
					"200": {
						"description": "Login successful",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/VerifyOTPResponse"
								}
							}
						}
					},
					"400": {
						"description": "Invalid OTP or login failed"
					}
				},
				"description": "Verify OTP and login user",
				"summary": "Verify login OTP and authenticate user",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/VerifyOTPRequest"
							}
						}
					}
				}
			}
		},
		"/api/auth/me": {
			"get": {
				"operationId": "GetCurrentUser",
				"responses": {
					"200": {
						"description": "User information retrieved",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					}
				},
				"description": "Get current authenticated user",
				"summary": "Get current user information",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": []
			}
		},
...
	},
	"info": {
		"title": "MotorQ API",
		"version": "1.0.0",
		"description": "API for the MotorQ application",
		"contact": {
			"name": "API Support",
			"email": "support@motorq.com"
		}
	},
	"servers": [
		{
			"url": "/",
			"description": "Default API server"
		}
	]
}